packages/parser/src/utils.py
"""
Utility functions and helpers for Paper2Data parser.

Includes file operations, text processing, logging setup, and validation.
"""

import logging
import sys
from pathlib import Path
from typing import Dict, Any, Optional, List, Union
import json
import yaml


def setup_logging(
    level: str = "INFO",
    log_file: Optional[Path] = None,
    format_string: Optional[str] = None
) -> logging.Logger:
    """Set up logging configuration for the parser."""
    # TODO: Implement logging setup
    logger = logging.getLogger("paper2data.parser")
    logger.info("Setting up logging configuration")
    raise NotImplementedError("Logging setup not yet implemented")


def validate_input(input_source: str) -> bool:
    """Validate input source (file path, URL, or DOI)."""
    # TODO: Implement input validation
    logging.info(f"Validating input: {input_source}")
    raise NotImplementedError("Input validation not yet implemented")


def format_output(
    data: Dict[str, Any],
    output_format: str = "json"
) -> Union[str, bytes]:
    """Format extracted data for output."""
    # TODO: Implement output formatting
    logging.info(f"Formatting output as {output_format}")
    raise NotImplementedError("Output formatting not yet implemented")


def clean_text(text: str) -> str:
    """Clean and normalize extracted text."""
    # TODO: Implement text cleaning
    logging.info("Cleaning extracted text")
    raise NotImplementedError("Text cleaning not yet implemented")


def extract_filename_from_url(url: str) -> str:
    """Extract a reasonable filename from a URL."""
    # TODO: Implement filename extraction
    logging.info(f"Extracting filename from URL: {url}")
    raise NotImplementedError("Filename extraction not yet implemented")


def ensure_directory(path: Path) -> Path:
    """Ensure directory exists, create if it doesn't."""
    # TODO: Implement directory creation
    logging.info(f"Ensuring directory exists: {path}")
    raise NotImplementedError("Directory creation not yet implemented")


def save_json(data: Dict[str, Any], filepath: Path) -> None:
    """Save data as JSON file."""
    # TODO: Implement JSON saving
    logging.info(f"Saving JSON to: {filepath}")
    raise NotImplementedError("JSON saving not yet implemented")


def load_config(config_path: Optional[Path] = None) -> Dict[str, Any]:
    """Load configuration from YAML file."""
    # TODO: Implement config loading
    logging.info(f"Loading configuration from: {config_path}")
    raise NotImplementedError("Config loading not yet implemented")


def get_file_hash(filepath: Path) -> str:
    """Generate hash for file content."""
    # TODO: Implement file hashing
    logging.info(f"Generating hash for: {filepath}")
    raise NotImplementedError("File hashing not yet implemented")


def progress_callback(current: int, total: int, message: str = "") -> None:
    """Callback function for progress tracking."""
    # TODO: Implement progress tracking
    if message:
        logging.info(f"Progress: {current}/{total} - {message}")
    else:
        logging.info(f"Progress: {current}/{total}")
    # This is a placeholder - real implementation would update progress bars


class ProcessingError(Exception):
    """Custom exception for processing errors."""
    pass


class ValidationError(Exception):
    """Custom exception for validation errors."""
    pass


class ConfigurationError(Exception):
    """Custom exception for configuration errors."""
    pass 